feat: integrate MUI components and enhance UI for image upload and detection
- Added MUI dependencies for improved UI components.
- Refactored App component to utilize MUI components for layout and controls.
- Implemented image overlay for displaying detection results.
- Created JSONTree component for rendering JSON data in a structured format.
- Enhanced logging functionality with buttons to view JSON data.
- Updated styles for better responsiveness and visual hierarchy.

+++ back/utils.py
@@ -90,10 +90,10 @@ class ImageAnalyzer:
                     {
                         "label": label,
                         "confidence": conf,
-                        "x1": float(x1),
-                        "y1": float(y1),
-                        "x2": float(x2),
-                        "y2": float(y2),
+                        "x1": int(x1),
+                        "y1": int(y1),
+                        "x2": int(x2),
+                        "y2": int(y2),
                     }
                 )
